% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpa_d2v_preprocess.R
\name{dpa_d2v_preprocess}
\alias{dpa_d2v_preprocess}
\title{Prepare a doc2vec matrix using a pretrained word2vec model.}
\usage{
dpa_d2v_preprocess(d2v.col, doc_id, text, as.df = T)
}
\arguments{
\item{doc_id}{the doc ids, probably the BID with Bastiat}

\item{text}{character vector of words separated by spaces}

\item{as.df}{return as a dataframe? if F, returns a matrix}

\item{model.w2v}{pretrained word embeddings}
}
\value{

}
\description{
Prepare a doc2vec matrix using a pretrained word2vec model.
}
