% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpa_generate_pr_metrics.R
\name{dpa_generate_pr_metrics}
\alias{dpa_generate_pr_metrics}
\title{Generates precision, and recall metrics for a given IR model. Takes in
vectors of the model's predictions, the real label for the documents and an
optional model name.}
\usage{
dpa_generate_pr_metrics(model.prediction, real.label, model.name = "model")
}
\arguments{
\item{model.prediction}{The prediction of your model, boolean.}

\item{real.label}{The real label of the document, boolean.}

\item{model.name}{The name of the model (optional)}
}
\value{
data.frame with the scores
}
\description{
Generates precision, and recall metrics for a given IR model. Takes in
vectors of the model's predictions, the real label for the documents and an
optional model name.
}
\references{
www.globaltradealert.org
}
