% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpa_update_library.R
\name{dpa_update_library}
\alias{dpa_update_library}
\title{Get the latest version from GitHub.}
\usage{
dpa_update_library(x)
}
\value{
Be up to date with our latest functions.
}
\description{
Syncs your local library with our latest GTA GitHub release.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
